requirements.dev.txt
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==43.0.1
mypy==1.11.2
mypy-extensions==1.0.0
pdfminer.six==20240706
pycparser==2.22
python-magic-bin==0.4.14 ; platform_system == 'Windows'
python-magic==0.4.27 ; platform_system != 'Windows'
tomli==2.0.1
typing_extensions==4.12.2
